web_search/requirements.txt
urllib3==1.26.18
unstructured==0.6.6
tabulate==0.9.0
pytesseract==0.3.10
tiktoken==0.4.0

# LangChain
langchain-google-vertexai==0.0.3
langchain==0.1.0
langchain-community==0.0.13

# Slack Bolt
slack-bolt==1.18.0

google-cloud-discoveryengine==0.11.8

# Load Web Page Documents
urllib3==1.26.18
beautifulsoup4==4.12.2
html2text==2020.1.16

# Vector DB
# chromadb==0.3.23
# faiss-gpu==1.7.2
google-cloud-bigquery==3.13.0

# Load Local file
# pdf2image==1.16.3
